Prefer `e.is_number` over `not e.free_symbols` in SymPy
We spend somewhere on the order 1% in `sympy.Expr.free_symbols` as it is called millions of times.
Most of the time we actually just want to know "is this a constant", however `e.is_constant()` is
horribly slow. It turns out though that there is another propery `is_number` that does what we want.

> property is_number:
>
> Returns True if self has no free symbols and no undefined functions (AppliedUndef, to be precise). It will be faster
> than if not self.free_symbols, however, since is_number will fail as soon as it hits a free symbol or undefined
> function.

Even further, we also avoid the overhead of building the unnecessary set object.

ghstack-source-id: e8b21881c89b69a48247b2c1bc948a777ca93664
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112688