[fx] Speedup ShapeEnv cache invalidation checks
This may seem a bit silly but we spend ~5% of compilation on simply checking if the `ShapeEnv` cache
has been invalidated. It isn't necessarily slow, but we call it millions of times per compile so
everything adds up.

To improve the situation, I've added a version counter to the shape env that gets incremented
whenever the cache key changes. This does require a bit of care in `ShapeEnv` that we don't modify
the relevant state without calling `self._update_version_counter()`. However, we already have a
similar situation for the translation validation feature which requires `_set_replacement` to be
called instead of modifying the replacements directly.

ghstack-source-id: 5402f795c35f180126695343f6e017333122b530
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112687