Deprecate "fallthrough" as autograd fallback default
This got reverted a couple of months ago. We have since fixed the known
problems with the flag. It is time to try again.

Context:

This PR adds a new fallback to the Autograd dispatch keys. The previous
behavior was a big footgun; we are moving to deprecating it.

If you would prefer the old behavior:
- A quick (unsupported) way to get the previous behavior is to call
torch._C._set_autograd_fallback("nothing")
- Register "torch::CppFunction::makeFallthrough()" to your Autograd key,
like in https://gist.github.com/zou3519/d09a5f4b1afe2430af09fea67c6ff2c8

It is possible that this PR regresses performance of overhead-bound
models. If this is the case, please reach out (and apply one of the
temporary fixes in the previous section).

Test Plan:
- existing tests

ghstack-source-id: 24b4ff176e146006c097d75368e6ce6340c89e78
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113166