CMake: Loosen CUDA consistency check
Closes #108931, closes #108932, see also conda-forge/pytorch-cpu#203

Currently we compare `CUDA_INCLUDE_DIRS` and expect exact equality
with `CUDAToolkit_INCLUDE_DIR` however this fails in the presense of
symbolic links or for split installs where there are multiple include paths.
Given that, it makes sense to loosen the requirement to just version
equality under the assumption that two installs of the same version
should still be compatible.

ghstack-source-id: de480550b7fd8b8c48e7854051c8097fd1e05c9d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113174