[dynamo] Avoid eager imports of classes with custom VariableTrackers
Currently custom VariableTrackers exist for classes that live outside of pytorch.
For these cases dynamo currently eagerly imports the module to get the Class
object to compare against.

This instead uses `sys.modules.get("module.path")` such that the module is never
imported by dynamo itself, but if the user has imported the module then we will
still access the module and grab the type we need to compare against.

I noticed this issue because importing `KeyedJaggedTensor` fails half-way
through if `fbgemm_gpu` has been built with an incompatible PyTorch version, in
which case it retries the import again each time!

ghstack-source-id: a77b1ca4d6259b7b89115737f79d1507ed9336c5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112319