[torchelastic] ensure grandchild processes are restarted correctly
When torchelastic notices that one rank has failed, it will sent a SIGTERM
signal to other trainer ranks to tear them down before restarting. However,
if the trainer itself launches subprocesses, or is launched by a non-python
wrapper script, then the SIGTERM will be delived only to the direct child of
torch eleastic and not all descendants. This opens subprocesses in a new
linux 'session' which starts a new process group with the pgid the same
as the trainers pid. Then when we send signals, we deliver them to the
process group rather than just the direct child.

ghstack-source-id: 5672a1dec17fb1de0a834d2bb5e63b4a9d30cc83
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113231