[dynamo] Make decorators.py pass follow-import typechecking
I am trying to turn on `follow_imports=silent` for MYPYNOFOLLOW.
However, this requires a huge number of changes, so I am breaking it
down to a per-file basis.

Unfortunately, we will not be able to turn on `follow_imports` until all
files are fixed, so there is no way to stop regressions. So I hope to get
these fixes in as fast as possible.

ghstack-source-id: 3e5b8b124a1fecc96e059e79d761aa661470d824
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113304