Make FakeProcessGroup traceable
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113314

This PR mimics what we have done to trace ProcessGroup. This allows use to use FakeProcessGroup with torch.compile. FakeProcessGroup allows us to use world_size > 1 without creating multiple processes thus enabling the usage of PDB to debug bucketing DDP allreduce in the Inductor. We can theoretically use GLOO with world_size==1 to achieve the same goal. However, the `wait()` seems to be optimized away when the world_size is 1.

ghstack-source-id: 206958093
@exported-using-ghexport

Differential Revision: [D51136463](https://our.internmc.facebook.com/intern/diff/D51136463/)